maj V5 avec page /populate pour remplir /export pour export par mail /group/match/add pour le html pour ajouter /group/match en post pour ajouter

+++ trunk/src/main/groovy/org/petank/server/dao/DAOManager.groovy
@@ -49,7 +49,7 @@ public class DAOManager {
 		def object
 		def PersistenceManager pm = initTransaction();
 		try {
-			object = pm.getObjectById(clazz, id)
+			object = pm.getObjectById(clazz, (id as Long))
 			commitTransaction()
 		} finally {
 			closeTransaction()
